Move Git settings to ~/.config/git/
This commit also changes my default settigs, add a commit template based
on a blog post by Drew DeVault[1], and add a global gitignore file.

There are many changes into one commit - ironically, one of the things
the template tells me not to do -, because this is a dotfile repository,
and I did not care enough to do it right. Sorry.

[1] https://drewdevault.com/2019/02/25/Using-git-with-discipline.html

Signed-off-by: James Pond <james@cipher.host>