Add my accounts configuration file for aerc
I was not going to add this here, but thinking about it, there is not
much that can be done with it, that cannot be done without. You can
already figure out my email provider by looking at my MX records, my
username is obvious, and the password is not located in the file.

Might as well keep the configuration file here, and make my life easier.

Signed-off-by: James Pond <james@cipher.host>