Change the filename length for cpimg
URLs are smaller now, similar to how YouTube, and Imgur handle things.
By using Argon2 in combination with a random pepper, and a hash of the
image name, file collision should not happen either.

This script is unnecessarily complicated, though, and I need to optimize
a few things, but it works.

Signed-off-by: James Pond <james@cipher.host>