Bump MPV version to 2.0.0
I was reviewing a video of a training session, and noticed that I needed
an option to delete the file after watching it. So I added it (in a
hacky way, should use trap instead). Use it --and-delete.

Then I realised I needed to access some options from MPV that this
script would not allow, so I rewrote it in a better way. Now if no
argument is passed to it, regular MPV takes over, while --background
leaves the video playing in the background.

Signed-off-by: James Pond <james@cipher.host>