Add initial version of the CLI Font Installer
This script should download, and install all fonts from Google Fonts in
its current state, but the idea here is to use it to install fonts from
user input, and whatnot.

I am not sure if I will ever get to finish this, but it solves my
problem for now, so I am happy.

Signed-off-by: James Pond <james@cipher.host>