Add gen-ssh-key, a wrapper that generates SSH keys
When I need to create an SSH key, I end up going through my bash history
to remember what are the defaults I use, and then enter a rabbit hole on
the internet to ensure these defaults are still current.

This is just an initial version to prevent me from going through that
process from time to time, but eventually I want to rewrite it, and make
it user-friendly, so Cipher Host customers can use it, since we do not
allow passwords for SSH, and SFTP over there.

The customer-friendly version should have a file that handles calling
gen-ssh-key, and then upload the .pub file to our encrypted paste
service, echoing back to the terminal a message that the customer can
send over to our support team, to add the file to the server.

Once the support team access the file on the paste service, it should
self-destroy. The file, not the team, haha.

Signed-off-by: James Pond <james@cipher.host>