Fix bug with cpimg when a GIF has too many colors
When I convert video recordings of my screen into GIFs for
documentation, gifsicle complains that the video has too many colors.

This change fixes the problem by reducing the number of colors in the
image when optimizing it. We could probably use --use-colormap web
instead, but I think --colors 256 is a nice balance here.

Signed-off-by: James Pond <james@cipher.host>