Add a wrapper for 1Password CLI with PINEntry
While I love the 1Password CLI, the fact that it just throw an error if
I try to get an item without having a valid session first, bugs the heck
out of me. And breaks my shell scripts too.

By integrating it with PINEntry, it now just asks for a password if my
session expired, allowing me to start a new session.

Signed-off-by: James Pond <james@cipher.host>