Add xsel wrapper to clear clipboard after X secs
I use xsel to copy passwords from 1Password, but leaving them in the
clipboard makes me uneasy, so this wrapper adds a way to copy to the
clipboard, wait a few seconds, and then clear it.

It is a very ugly hack, but it works, which is all that matters in the
end.

Signed-off-by: James Pond <james@cipher.host>