[unlock-n-mount] Add YubiKey as a requirement
I am not a security expert, so take this with a grain of salt, but the
idea here is that an attacker now needs to breach multiple layers to get
to the content of the encrypted device.

Assuming the attacker knows your login password, they would need:

  1. Physical access to the hard drive.
  2. Physical access to the YubiKey.
  3. Physical access to the computer with the password store.
  4. Physical access to the phone with the 1Password 2FA codes.
  5. The password for the GPG key used to encrypt the password store.
  6. The master password for the 1Password account.
  7. The 2FA code to access the 1Password account.

While more secure than before in my opinion[1], these changes do add a
non-insignificant amount of time to the unlocking time and make the
process less convenient and more expensive.

Alternatively, the attacker needs a wrench[2].

[1] Please correct me if I am wrong, email address is a few lines down.
[2] https://xkcd.com/538/

Signed-off-by: James Pond <james@cipher.host>