[gotrng] Remove rngd as the feeding method
According to the OneRNG developers, rng-tools is broken with Linux
kernel 5.6 and higher, so I am removing it from the tool as I have no
server or workstation running anything older than 5.8.

I still need to read more about this as I see no proof that the tool is
indeed broken with newer versions of the kernel, though.

Signed-off-by: James Pond <james@cipher.host>